refactor: CardValidation 에러 상태 업데이트 로직 분리
- 중복 코드 통합
